[ISSUES 135] App can not start up with RocketMQ-streams  (#136)
* fix(jar) make rocketmq-streams runnable as a sdk

* fix(example)make example runnable
